Add DescriptorTransformer; update ByValue

+++ src/main/java/overrun/marshal/Checks.java
@@ -32,7 +32,7 @@ public final class Checks {
      * @param expected the expected size
      * @param actual   the actual size
      */
-    public static void checkArraySize(int expected, int actual) {
+    public static void checkArraySize(long expected, int actual) {
         if (MarshalConfigs.CHECK_ARRAY_SIZE.get() && expected != actual) {
             throw new IllegalArgumentException("Expected array of size " + expected + ", got " + actual);
         }
